% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.SummarizedExperiment}
\title{create data.frame}
\usage{
\method{as.data.frame}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{\code{SummarizedExperiment} object}

\item{...}{unused arguments}
}
\description{
create data.frame
}
