% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-group-.R
\name{bioc_group_id}
\alias{bioc_group_id}
\alias{bioc_group_loc}
\title{Recycle a vector}
\usage{
bioc_group_id(x, ...)

bioc_group_loc(x, ...)
}
\arguments{
\item{x}{A vector}
}
\value{
either a tibble or DataFrame object
}
\description{
A alternative to \code{\link[vctrs:vec_group]{vctrs::vec_group_id}} and
\code{\link[vctrs:vec_group]{vctrs::vec_group_loc}} as
an S4 generic function to allow \code{S4Vectors}.
}
\examples{

data <- DataFrame(
  letter = sample(letters, 500, TRUE),
  LETTER = sample(letters, 500, TRUE)
)

bioc_group_id(data)
bioc_group_loc(data)

}
