% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-init.r
\name{bioc_init}
\alias{bioc_init}
\title{bioc_init}
\usage{
bioc_init(x, size = 0L, ...)
}
\arguments{
\item{x}{the prototype object. typically unused}

\item{size}{the size of the object to create}

\item{...}{unused}
}
\description{
create an object of some size. This generic does not have an equivalent
function from the vctrs pacakge however it was created to support the
funcitonality of \code{bioc_cast()} between two data.frame-like objects.

The default behavior of \code{vctrs::vec_cast(x, to)} when x is a data.frame and
and \code{to} is another, potentially bigger data.frame, is to fill in the missing
names with missing prototype vectors.

The vctrs pacakge provides a way to potentially cover any arbitrary case if
both vctrs::vec_proxy and vctrs::vec_restore are defined for the class.

However, this is not always the case for S4Vectors::Vectors and other S4
classes. Thus, \code{bioc_init()} was created to provide developers a way to
create another default instance of their classes if dispatched upon.
}
