% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-ptype.R, R/bioc-ptype2.R
\name{bioc_ptype}
\alias{bioc_ptype}
\title{create a prototype of an object}
\usage{
bioc_ptype(x, ...)

bioc_ptype2(x, y, ...)
}
\value{
an object of size 0

an object of size 0
}
\description{
get the zero size prototype of an object

get the zero size prototype of an object
}
