% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-unchop.r
\name{bioc_unchop}
\alias{bioc_unchop}
\alias{bioc_unchop_ptype}
\title{unchop a list-like object into some prototype}
\usage{
bioc_unchop(x, ptype, ..., indices = NULL)

bioc_unchop_ptype(ptype, x, ..., indices = NULL)
}
\arguments{
\item{x}{a list}

\item{ptype}{prototype to unchop into}

\item{...}{unused}

\item{indices}{optional list of integer vectors giving locations to splice
elements x into the final vector.}
}
\description{
unchop a list-like object into some prototype
}
