% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocmask-R6.R
\name{biocmask_assay}
\alias{biocmask_assay}
\title{\code{biocmask} for SummarizedExperiment \code{assays()}}
\description{
A more specialized version of the biocmask R6 object for the
assays list object. This includes chopping and unchopping
of matrix like objects.
}
\section{Super class}{
\code{\link[biocmask:biocmask]{biocmask::biocmask}} -> \code{biocmask_assay}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-biocmask_assay-new}{\code{biocmask_assay$new()}}
\item \href{#method-biocmask_assay-eval}{\code{biocmask_assay$eval()}}
\item \href{#method-biocmask_assay-unchop}{\code{biocmask_assay$unchop()}}
\item \href{#method-biocmask_assay-clone}{\code{biocmask_assay$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biocmask" data-topic="biocmask" data-id="bind"><a href='../../biocmask/html/biocmask.html#method-biocmask-bind'><code>biocmask::biocmask$bind()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biocmask" data-topic="biocmask" data-id="on_bind"><a href='../../biocmask/html/biocmask.html#method-biocmask-on_bind'><code>biocmask::biocmask$on_bind()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biocmask" data-topic="biocmask" data-id="results"><a href='../../biocmask/html/biocmask.html#method-biocmask-results'><code>biocmask::biocmask$results()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biocmask_assay-new"></a>}}
\if{latex}{\out{\hypertarget{method-biocmask_assay-new}{}}}
\subsection{Method \code{new()}}{
Create a biocmask from \code{.data}. \code{.data} is chopped by
\code{.indices}, and environments are built from \code{.env}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biocmask_assay$new(.data, .indices, .env, .nrow, .ncol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{a named list like object to create a mask}

\item{\code{.indices}}{the indices that will be used to chop \code{.data}}

\item{\code{.env}}{an environment that the resulting mask will be built from.}

\item{\code{.nrow, .ncol}}{the number of rows and columns of each element of \code{.data} respectively}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biocmask_assay-eval"></a>}}
\if{latex}{\out{\hypertarget{method-biocmask_assay-eval}{}}}
\subsection{Method \code{eval()}}{
evaluates a quoted expression within a new datamask, forces results into
a matrix of the expected size for the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biocmask_assay$eval(quo, env = caller_env())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quo}}{a quosure to evaluate}

\item{\code{env}}{an environment to search after mask}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biocmask_assay-unchop"></a>}}
\if{latex}{\out{\hypertarget{method-biocmask_assay-unchop}{}}}
\subsection{Method \code{unchop()}}{
unchop data within the mask, returns a matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biocmask_assay$unchop(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of binding to retrieve and unchop}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-biocmask_assay-clone"></a>}}
\if{latex}{\out{\hypertarget{method-biocmask_assay-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{biocmask_assay$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
