% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context-utils.R
\name{biocmask-context}
\alias{biocmask-context}
\alias{cols}
\alias{rows}
\alias{col_ctx}
\alias{row_ctx}
\alias{assay_ctx}
\title{biocmask contexts}
\usage{
cols(...)

rows(...)

col_ctx(x)

row_ctx(x)

assay_ctx(x)
}
\arguments{
\item{...}{expressions to evaluate within its associated context}
}
\description{
Contextual user-facing helper function for dplyr verbs with SummarizedExperiment
objects. These functions are intended to be used as the top level call to
any dplyr verbs \code{...} argument, similar to that of \code{across()}/\code{if_any()}/\code{if_all()}.

Specifies that the following expressions should be evaluated within the
colData context.

Specifies that the following expressions should be evaluated within the
rowData context.

Specify a single expression to evaluate in another context

Specify a single expression to evaluate in another context

Specify a single expression to evaluate in another context
}
