% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocmask-package.R
\docType{package}
\name{biocmask-package}
\alias{biocmask-package}
\title{biocmask: Data masks utilities for S4 objects using S7 functions inspired by vctrs and rlang.}
\value{
API for using S4 classes with rlang data masks
}
\description{
The package provides `rlang`-like data masks for the complex class objects. The enables the evaluation of unquoted expression in different contexts of the object with optional access to other contexts. The goal for `biocmask` is for evaluation to feel like a data.frame object without ever needing to unwind to a rectangular data.frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jtlandis/biocmask}
  \item \url{https://jtlandis.github.io/biocmask}
  \item Report bugs at \url{https://www.github.com/jtlandis/biocmask/issues}
}

}
\author{
\strong{Maintainer}: Justin Landis \email{jtlandis314@gmail.com} (\href{https://orcid.org/0000-0001-5501-4934}{ORCID})

Authors:
\itemize{
  \item Michael Love \email{michaelisaiahlove@gmail.com} (\href{https://orcid.org/0000-0001-8401-0545}{ORCID})
}

}
\keyword{internal}
