% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{vec_phantom}
\alias{vec_phantom}
\alias{biocmask-printing}
\alias{biocmask_pillar_format}
\title{Printing within tibble with S4 objects}
\usage{
vec_phantom(x)

biocmask_pillar_format(x, ...)
}
\arguments{
\item{x}{The S4 object}

\item{...}{other arguments passed from \code{\link[pillar:pillar_shaft]{pillar_shaft}}}
}
\value{
\code{biocmask_pillar_format} -> formatted version of your S4 vector
\code{vec_phantom} -> integer vector with arbitrary object in \code{phatomData} attribute.
}
\description{
\code{biocmask} uses \link[pillar:pillar-package]{pillar} for its printing.
If you want to change how your S4 object is printed within
\code{biocmask}'s print method, consider writing a method for
this function.

To print S4 objects in a tibble, \code{biocmask} hacks a custom
integer vector built from \code{\link[vctrs:new_vctr]{vctrs}} where
the S4 object lives in an attribute named "phantomData".
You can create your own S4 phantom vector with \code{vec_phantom()}.
This function is not used outside of printing for \code{biocmask}

The default method for formatting a \code{vec_phantom()} is to call
\code{\link[S4Vectors:show-utils]{showAsCell()}}.
}
