% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_quos.R
\name{biocmask_quos}
\alias{biocmask_quos}
\title{biocmask quosures}
\usage{
biocmask_quos(..., .named = TRUE, .ctx_default = NULL, .ctx_opt = NULL)
}
\arguments{
\item{...}{rlang dots, supports splicing an quoting}

\item{.named}{should resulting expressions be named?}

\item{.ctx_default}{default context to eval within}

\item{.ctx_opt}{optional contexts to eval within}
}
\value{
a quosure with attribute \code{biocmask:::ctx}.
}
\description{
a consistent way to handle \code{...} for dplyr extensions.
This returns a list of quosures where each quosure
contains an attribute \code{biocmask:::ctx} indicating which
mask context it should be evaluate in.
}
\examples{

# in biocmask the default context is "assays"
# and optional contexts are "rows" and "cols"
quos <- biocmask_quos(
  foo = bar,
  ctx2(foo = bar),
  ctx3(foo = bar),
  .ctx_default = "ctx1",
  .ctx_opt = c("ctx2", "ctx3")
)
attr(quos[[1]], "biocmask:::ctx")
attr(quos[[2]], "biocmask:::ctx")
attr(quos[[3]], "biocmask:::ctx")

}
