% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-S4-.R
\docType{data}
\name{class_vctrs}
\alias{class_vctrs}
\title{S7 classes for vctrs and S4 Vectors
useful for S7 method dispatch}
\format{
An object of class \code{S7_union} of length 1.
}
\usage{
class_vctrs
}
\description{
S7 classes for vctrs and S4 Vectors
useful for S7 method dispatch
}
\keyword{datasets}
