% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_simple}
\alias{df_simple}
\title{Biocmask Simple Example Summarized Experiment}
\format{
\subsection{\code{df_simple}}{

\describe{
\item{\code{signal}}{sampled data points between 1:100}
\item{condition}{an Rle object of "cntrl" and "drug"}
\item{group}{a factor vector of "A", "B", "C"}
}
}
}
\usage{
df_simple
}
\value{
a \code{SummarizedExperiment} object
}
\description{
A small DataFrame Object of 8 observations, and 3 columns
}
\examples{
df_simple |> names()
}
\keyword{datasets}
