% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/across.R
\name{dplyr_across_internals}
\alias{dplyr_across_internals}
\alias{expr_substitute}
\alias{node_walk_replace}
\alias{as_across_expr}
\alias{is_inlinable_lambda}
\alias{validate_fns}
\alias{is_inlinable_function}
\alias{is_inlinable_formula}
\title{dplyr across internals}
\usage{
expr_substitute(expr, old, new)

node_walk_replace(node, old, new)

as_across_expr(fn, var)

is_inlinable_lambda(x)

validate_fns(quo, mask, error_call = caller_env())

is_inlinable_function(x)

is_inlinable_formula(x)
}
\description{
the following functions were taken verbatim from the \code{dplyr} package such
that behaviors of \code{across()} were consistent between \code{dplyr} and \code{biocmask}
}
\keyword{internal}
