% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.SummarizedExperiment}
\alias{filter.SummarizedExperiment}
\title{Filter SummarizedExperiment}
\usage{
\method{filter}{SummarizedExperiment}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A SummarizedExperiment object}

\item{...}{conditions to filter on. These must be wrapped in \code{cols()} and or
\code{rows()}}

\item{.preserve}{Relevant when the .data input is grouped. If .preserve = FALSE
(the default), the grouping structure is recalculated based on the resulting data,
i.e. the number of groups may change.}
}
\value{
SummarizedExperiment Object
}
\description{
Filtering operations can only return a SummarizedExperiment
object so long as they are performed on rows and cols contexts.
This is because filtering on assays does not gaurentee a recoverable
SummarizedExperiment structure.
}
