% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by}
\alias{group_by}
\alias{group_by.SummarizedExperiment}
\alias{ungroup.SummarizedExperiment}
\title{group_by SummarizedExperiment}
\usage{
\method{group_by}{SummarizedExperiment}(.data, ..., .add = FALSE)

\method{ungroup}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{.data}{a SummarizedExperiment object}

\item{...}{}

\item{.add}{When \code{FALSE}, the default, \code{group_by()} will override
existing groups.}

\item{x}{A SummarizedExperiment object}
}
\value{
SummarizedExperiment object
}
\description{
create grouping variables about the rowData and colData of a
SummarizedExperiment object. Unlike the \code{data.frame} method
the resulting output class is left unchanged. Thus \code{dplyr} generics for
\code{SummarizedExperiment} must check grouping information manually.

Ungroup a SummarizedExperiment object
}
