% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-group_by.R
\name{group_by.SummarizedExperiment}
\alias{group_by.SummarizedExperiment}
\alias{ungroup.SummarizedExperiment}
\title{apply groups to SummarizedExperiments}
\usage{
\method{group_by}{SummarizedExperiment}(.data, ..., .add = FALSE)

\method{ungroup}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{.data}{a SummarizedExperiment object}

\item{...}{\link[=biocmask-context]{contextual expressions} specifying
which columns to ungroup. Omitting \code{...} ungroups the entire object.}

\item{.add}{When \code{FALSE}, the default, \code{group_by()} will override
existing groups.}

\item{x}{A SummarizedExperiment object}
}
\value{
SummarizedExperiment object
}
\description{
create grouping variables about the rowData and colData of a
SummarizedExperiment object. Unlike the \code{data.frame} method
the resulting output class is left unchanged. Thus \code{dplyr} generics for
\code{SummarizedExperiment} must check grouping information manually.
}
\section{Functions}{
\itemize{
\item \code{ungroup(SummarizedExperiment)}: Ungroup a SummarizedExperiment object

}}
\examples{

group_by(se_simple, rows(direction), cols(condition))

}
