% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_data-SummarizedExperiment}
\alias{group_data-SummarizedExperiment}
\alias{group_data.SummarizedExperiment}
\title{get grouping data}
\usage{
\method{group_data}{SummarizedExperiment}(.data)
}
\arguments{
\item{.data}{an object}
}
\description{
retrieve grouping information from a \code{SummarizedExperiment} object. This
is stored within the \code{metadata()} of the object.
}
