% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\alias{mutate.SummarizedExperiment}
\title{Mutate a SummarizedExperiment object}
\usage{
\method{mutate}{SummarizedExperiment}(.data, ...)
}
\arguments{
\item{.data}{a SummarizedExperiment object}

\item{...}{expressions}
}
\value{
SummarizedExperiment object
}
\description{
Mutate a SummarizedExperiment object under an data mask
}
