% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_env_top.R
\name{new_bioc_bot_env}
\alias{new_bioc_bot_env}
\title{Create new environment prior to biocmask.}
\usage{
new_bioc_bot_env(..., context = NULL, parent = top_env)
}
\arguments{
\item{context}{The context to use for the new environment.}

\item{parent}{The parent environment to inherit from. Ideally this is the #' environment created by \code{new_bioc_top_env()}.}
}
\description{
dictate the default context as well as the group id.
}
