% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_env_top.R
\name{new_bioc_top_env}
\alias{new_bioc_top_env}
\title{Create a new top-level environment for data masking.}
\usage{
new_bioc_top_env(..., parent = empty_env())
}
\arguments{
\item{...}{Additional arguments to pass to \code{new_environment()}.}

\item{parent}{The parent environment to inherit from.}
}
\description{
Creates a new environment that inherits from the specified parent
environment. This environment is automatically installed with various
\verb{bioc_*} functions that may be useful for the user.

Ideally, this environment should be created once within a package's
namespace, and then biocmasks should construct new environments with
this top level as the parent environment.
}
