% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocmask.R
\name{new_biocmask_manager}
\alias{new_biocmask_manager}
\alias{new_biocmask_manager.SummarizedExperiment}
\title{New Biocmask}
\usage{
new_biocmask_manager(obj, ...)

\method{new_biocmask_manager}{SummarizedExperiment}(obj, ...)
}
\arguments{
\item{obj}{Dispatch Object}

\item{...}{Not used}
}
\value{
a biocmask_manager R6 class object
}
\description{
Create a biocmask for an object
}
\examples{

manager <- new_biocmask_manager(se_simple)
manager$ctx
q <- biocmask_quos(
  counts_1 = counts + 1,
  cols(is_drug = condition == "drug"),
  .ctx_default = "assays",
  .ctx_opt = c("rows", "cols")
)
manager$eval(q[[1]])
manager$results()
# evaluating second quo without switching contexts will error
manager$eval(q[[2]]) |> try()
manager$ctx <- "cols"
manager$ctx
manager$eval(q[[2]])
manager$results()

}
\seealso{
\link{BiocmaskManager}
}
