% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{new_view_spec}
\alias{new_view_spec}
\title{Specifying views}
\usage{
new_view_spec(
  ctx,
  from,
  ...,
  mapper = ~.x,
  asis = ~.subset2(.x, 1L),
  reshape = ~.x
)
}
\arguments{
\item{ctx}{character name of the source context}

\item{from}{character name of context we are viewing \code{ctx} from}

\item{...}{additional bindings from the \code{from} context that should be
made available to the new \code{view}.}

\item{mapper}{a mapping of \code{from}'s index to the corresponding chop index
(or indices) in \code{ctx}. use the \code{.mapper} variable to retrieve the indices
for \code{ctx}. the mapper function is bound to the \code{from}'s bottom environment.}

\item{asis}{an expression on how \code{ctx} should be viewed from the \code{from}
context. The \code{asis} expression is lazily evaluated in \code{ctx}'s chop
environment, but bound to an environment that is a child of \code{from}'s bottom
enviornment. note: \code{.x} will be replaced dynamically with the correct name.
Only symbols found within the \code{ctx}'s mask will be evaluated, thus the
developer is responsible for how dynamic these bindings are.}

\item{reshape}{an expression on how to change \code{asis} to fit the \code{from} context}
}
\description{
Specifying views
}
