% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate}
\alias{filter}
\alias{summarise}
\alias{group_by}
\alias{group_data}
\alias{group_vars}
\alias{ungroup}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_data}}, \code{\link[dplyr:group_data]{group_vars}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:group_by]{ungroup}}}
}}

