% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se_simple}
\alias{se_simple}
\title{Biocmask Simple Example Summarized Experiment}
\format{
\subsection{\code{se_simple}}{

\describe{

\item{\code{assays}}{
\describe{
\item{counts}{sampled data points between 1:20}
\item{logcounts}{log transform of \code{counts}}
}
}
\item{\code{rowData}/\code{.features}}{
\describe{
\item{gene}{fake gene name}
\item{length}{fake gene length}
\item{direction}{fake strand}
}
}
\item{\code{colData}/\code{.samples}}{
\describe{
\item{sample}{fake sample name}
\item{condition}{control or drug treatment}
}
}
}
}
}
\usage{
se_simple
}
\value{
a \code{SummarizedExperiment} object
}
\description{
A small data SummarizedExperiment Object of 20 observations, 5 rows and 4
columns.
}
\examples{
SummarizedExperiment::assays(se_simple)
SummarizedExperiment::rowData(se_simple)
SummarizedExperiment::colData(se_simple)
}
\keyword{datasets}
