% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.SummarizedExperiment}
\alias{select.SummarizedExperiment}
\title{select assays, rowData, and colData names}
\usage{
\method{select}{SummarizedExperiment}(.data, ...)
}
\arguments{
\item{.data}{a \code{SummarizedExperiment} object}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> one or more selection
expressions. Supports the \code{SummarizedExperiment-context-functions}.}
}
\description{
Select one or more values from each context. By default omitting an expression
for a context is the same as selecting NOTHING from that context.
}
