% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-recycle.R
\name{vctrs-bioc_recycle}
\alias{vctrs-bioc_recycle}
\alias{bioc_recycle}
\title{Recycle a vector}
\usage{
bioc_recycle(x, size, ...)
}
\arguments{
\item{x}{A vector to recycle.}

\item{size}{Desired output size.}

\item{...}{Depending on the function used:
\itemize{
\item For \code{vec_recycle_common()}, vectors to recycle.
\item For \code{vec_recycle()}, these dots should be empty.
}}
}
\value{
a S3 or S4 vector
}
\description{
A alternative to \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle}} as an S4 generic
function to allow \code{S4Vectors}.
}
\examples{
bioc_recycle(1L, size = 5L)
bioc_recycle(S4Vectors::Rle(1L), size = 5L)

}
