% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-rep.R
\name{vctrs-bioc_rep}
\alias{vctrs-bioc_rep}
\alias{bioc_rep_each}
\title{replicate a vector}
\usage{
bioc_rep(x, times, ...)

bioc_rep_each(x, times, ...)
}
\arguments{
\item{x}{A vector.}

\item{times}{For \code{vec_rep()}, a single integer for the number of times to repeat
the entire vector.

For \code{vec_rep_each()}, an integer vector of the number of times to repeat
each element of \code{x}. \code{times} will be \link[vctrs:theory-faq-recycling]{recycled} to
the size of \code{x}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a new S3 or S4 vector replicated by specified times
}
\description{
A wrapper of \code{\link[vctrs:vec-rep]{vctrs::vec_rep}} and
\code{\link[vctrs:vec-rep]{vctrs::vec_rep_each}} as an S7 generic
function to allow \code{S4Vectors}.
}
\examples{
bioc_rep(1:2, times = 5)
bioc_rep(S4Vectors::Rle(1:2), times = 5)

bioc_rep_each(1:2, times = 5)
bioc_rep_each(S4Vectors::Rle(1:2), times = 5)
}
