% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-ptype2.R
\name{bioc_c}
\alias{bioc_c}
\title{concatinate objects}
\usage{
bioc_c(...)
}
\description{
an alternative to \code{base::c} and \code{vctrs::vec_c} that will
use \code{biocmask::bioc_unchop} under the hood.
}
