% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocmask-package.R
\docType{package}
\name{biocmask-package}
\alias{biocmask-package}
\title{biocmask: Create data masks for SE dplyr functions}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Justin Landis \email{jtlandis314@gmail.com} (\href{https://orcid.org/0000-0001-5501-4934}{ORCID})

}
\keyword{internal}
