% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{biocmask-printing}
\alias{biocmask-printing}
\alias{biocmask_pillar_format}
\title{Printing within tibble with S4 objects}
\usage{
biocmask_pillar_format(x, ...)
}
\arguments{
\item{x}{The S4 object}

\item{...}{other arguments passed from \code{\link[pillar:pillar_shaft]{pillar_shaft}}}
}
\description{
\code{biocmask} uses \link[pillar:pillar-package]{pillar} for its printing.
If you want to change how your S4 object is printed within
\code{biocmask}'s print method, consider writing a method for
this function.

To print S4 objects in a tibble, \code{biocmask} hacks a custom
integer vector built from \code{\link[vctrs:new_vctr]{vctrs}} where
the S4 object lives in an attribute named "phantomData".

The default method for formatting a \code{vec_phantom} is to call
\code{\link[S4Vectors:show-utils]{showAsCell()}}.
}
