% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocmask.R
\name{biocmask}
\alias{biocmask}
\alias{new_biocmask}
\alias{new_biocmask.SummarizedExperiment}
\title{New Biocmask}
\usage{
new_biocmask(obj, ...)

\method{new_biocmask}{SummarizedExperiment}(obj, ...)
}
\arguments{
\item{obj}{Dispatch Object}

\item{...}{Not used}
}
\description{
Create a biocmask for an object
}
