% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_quos.R
\name{biocmask_quos}
\alias{biocmask_quos}
\title{biocmask quosures}
\usage{
biocmask_quos(..., .named = TRUE, .ctx = NULL, .trans = list())
}
\arguments{
\item{...}{rlang dots, supports splicing an quoting}

\item{.named}{should resulting expressions be named?}

\item{.ctx}{A character string of availablecontexts. The first
element is considered to be the default context, whereas the
rest are considered optional.}

\item{.ctx_trans}{A list of functions to eventually apply to
the expressions depending on their context. The names of the list
elements should match the contexts names provided by \code{.ctx}. These
functions will be quoted and in-lined into the expressions and thus
cannot rely on the environments the functions are constructed.
These functions are quoted}
}
\value{
a quosure with attribute \code{biocmask:::ctx}.
}
\description{
a consistent way to handle \code{...} for dplyr extensions.
This returns a list of quosures where each quosure
contains an attribute \code{biocmask:::ctx} indicating which
mask context it should be evaluate in.
}
\examples{

quos <- biocmask_quos(
  foo = bar,
  ctx2(foo = bar),
  ctx3(foo = bar),
  .ctx_default = "ctx1",
  .ctx_opt = c("ctx2", "ctx3")
)
attr(quos[[1]], "biocmask:::ctx")
attr(quos[[2]], "biocmask:::ctx")
attr(quos[[3]], "biocmask:::ctx")

}
