% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by.SummarizedExperiment}
\alias{group_by.SummarizedExperiment}
\title{group_by SummarizedExperiment}
\usage{
\method{group_by}{SummarizedExperiment}(.data, ..., .add = FALSE)
}
\arguments{
\item{.data}{a SummarizedExperiment object}

\item{...}{expressions to group on. Grouping may only be done on
rowData and/or colData by \code{rows()} and \code{cols()} respectively.}

\item{.add}{When \code{FALSE}, the default, \code{group_by()} will override
existing groups.}
}
\value{
SummarizedExperiment object
}
\description{
create grouping variables about the rowData and colData of a
SummarizedExperiment object. Unlike the \code{data.frame} method
the resulting output class is left unchanged. Thus \code{dplyr} generics for
\code{SummarizedExperiment} must check grouping information manually.
}
\examples{
group_by(se_simple, rows(direction), cols(condition))

}
