% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-group_by.R
\name{group_data-SummarizedExperiment}
\alias{group_data-SummarizedExperiment}
\alias{group_data.SummarizedExperiment}
\title{get grouping data}
\usage{
\method{group_data}{SummarizedExperiment}(.data)
}
\arguments{
\item{.data}{an object}
}
\value{
list of groupings for an SummarizedExperiment
}
\description{
retrieve grouping information from a \code{SummarizedExperiment} object. This
is stored within the \code{metadata()} of the object.
}
\examples{
group_by(se_simple, rows(direction), cols(condition)) |> group_data()
}
