% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-SummarizedExperiment.R,
%   R/reexports-dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{SummarizedExperiment}
\alias{assay}
\alias{assay<-}
\alias{assays}
\alias{assays<-}
\alias{rowData}
\alias{rowData<-}
\alias{colData}
\alias{colData<-}
\alias{metadata}
\alias{metadata<-}
\alias{arrange}
\alias{mutate}
\alias{filter}
\alias{summarise}
\alias{select}
\alias{group_by}
\alias{group_data}
\alias{group_vars}
\alias{ungroup}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_data}}, \code{\link[dplyr:group_data]{group_vars}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{S4Vectors}{\code{\link[S4Vectors:Annotated-class]{metadata}}, \code{\link[S4Vectors:Annotated-class]{metadata<-}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData<-}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData<-}}, \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}}}
}}

