% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise.SummarizedExperiment}
\alias{summarise.SummarizedExperiment}
\title{Summarize SummarizedExperiment}
\usage{
\method{summarise}{SummarizedExperiment}(.data, ..., .retain = TRUE)
}
\arguments{
\item{.data}{a SummarizedExperiment object,}

\item{...}{expressions to summarize the object}

\item{.retain}{logical value. When TRUE (the default), ungrouped dimensions
are retained in the resulting SummarizedExperiment object and scalar outputs
are recycled to the length of the ungrouped dimension. When FALSE, all
outputs are expected to be scalar values and all columns in ungrouped
dimensions are dropped.}
}
\description{
Summarize SummarizedExperiment
}
\examples{

# outputs in assay context may be either
# length 1, or the length of the ungrouped
# dimension while .retain = TRUE
se_simple |>
  group_by(rows(direction)) |>
  summarise(col_sums = colSums(counts),
            sample = sample(1:20, 1L))

# .retain = FALSE will drop ungrouped dimensions and
# outputs of assay context should be length 1.
se_simple |>
  group_by(rows(direction)) |>
  summarise(col_sums = list(colSums(counts)),
            .retain = FALSE) 

# using an `across()` function will help
# nest ungrouped dimensions
se_simple |>
  group_by(rows(direction)) |>
  summarise(col_sums = list(colSums(counts)),
            cols(across(everything(), list)),
             .retain = FALSE)
 
}
