% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise.SummarizedExperiment}
\alias{summarise.SummarizedExperiment}
\title{Summarize SummarizedExperiment}
\usage{
\method{summarise}{SummarizedExperiment}(.data, ...)
}
\description{
Summarize SummarizedExperiment
}
